requirements.txt
numpy>=1.20
matplotlib>=3.5
Pillow>=9.0